+++ client/src/pages/CatalogPage/DeepDive/SearchTool/SearchTool.tsx
@@ -7,8 +7,8 @@ import { SelectValue } from 'antd/lib/select';
 import axios from 'axios';
 import { equalityDefault } from 'components/Forms/EqualityInput/EqualityInput';
 import { RangePickerValue } from 'antd/lib/date-picker/interface';
-import ParticipantForm from '../../ParticipantForm/ParticipantForm';
-import Participant from 'pages/CatalogPage/ParticipantForm/Participant';
+import ParticipantForm from '../ParticipantForm/ParticipantForm';
+import Participant from 'pages/CatalogPage/DeepDive/ParticipantForm/Participant';
 import { ANY_OPTION } from '../AnyOption';
 import { GunCrime } from 'pages/CatalogPage/GunCrime';
 import styles from './SearchTool.module.less';
